fix: rename DetailProduct component and update imports
