+++ src/components/productCatalog/DetailProduct.tsx
